Progress Manager
Added in saving unlocked levels.
Connected to the ActManager so that it saves when you complete an act and unlocks the next level.
Progress saves when you quit the game or win the act.
Progress loads when you start the game.
There is also a debug gui to save, load, and reset data.